add about section to website
