+++ pages/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useRef, useState } from "react";
 
+import About from "@/components/About";
 import Disrupt_Reality_Button from "@/components/Disrupt_Reality_Button";
 import { Events } from "react-scroll";
 import Image from "next/image";
@@ -193,11 +194,7 @@ export default function Home() {
           </div>
         </div>
 
-        <div className="laptop-wrapper">
-          <div className="laptop-container">
-            <Image src={laptopSvg} alt="Laptop" />
-          </div>
-        </div>
+        <About />
       </main>
     </>
   );
